DELIA-5466: Added log4c linking to rdk-loggers shared library.
Reason for change: since rdk-loggers is based on log4c,
we are adding a required library to the lib header,
without which anybody using rdkloggers would have to
link log4c along with it.

Each of the shared library needs to have its dependencies listed in the lib header
so the run time resolves the undefined symbols quickly.

Test Procedure: Make sure it is built and we have all the logs updated properly.

Risks: None

Change-Id: I7a6c00f4d2f98414c8a3b512396db5023b5445b8
Signed-off-by: Balaji Punnuru <balaji_punnuru@cable.comcast.com>